server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
paho-mqtt>=2.0.0
pydantic>=2.0.0
